organize imports to use globals, change folder structure
